gave permissions to managing director to read anything

+++ src/main/java/vector/StockManagement/model/dto/TransferDTO.java
@@ -7,8 +7,8 @@ public class TransferDTO {
 
     private Long oderId;
     private String reason;
-    private Long quantityToTransfer;
-    private Long quantityDelivered;
+    private Integer quantityToTransfer;
+    private Integer quantityDelivered;
     private Long fromId;
     private Long toId;
